Refactored org.exolab.castor.builder.types package (CASTOR-1813)

+++ castor/trunk/src/main/java/org/exolab/castor/xml/validators/IntegerValidator.java
@@ -228,7 +228,7 @@ public class IntegerValidator extends PatternValidator implements TypeValidator 
      * @param maxValue
      *            the maximum (exclusive) value for xsd:integer validation.
      */
-    public void setMaxExclusive(final int maxValue) {
+    public void setMaxExclusive(final long maxValue) {
         _useMax = true;
         _max = maxValue - 1;
     } // -- setMaxExclusive
